+++ docs/conf.py
@@ -4,7 +4,7 @@
 project = "DataLogger"
 copyright = "2023, California Institute of Technology"
 author = "Alex Hadley"
-release = "0.1.1"
+release = "0.2.0"
 
 # General configuration
 extensions = [